requirements.txt
bcrypt==3.2.0
capstone==4.0.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
idna==2.10
intervaltree==3.1.0
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.20.1
packaging==20.9
paramiko==2.7.2
Pillow==8.1.0
psutil==5.8.0
pycparser==2.20
pycryptodome==3.10.1
pyelftools==0.27
Pygments==2.8.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.25.1
ROPGadget==6.5
six==1.15.0
sortedcontainers==2.3.0
unicorn==1.0.2rc3
urllib3==1.26.3
